Refactor majority and class migrations to use appropriate data types and create partners table

+++ database/migrations/2025_03_20_021610_create_majorities_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('majorities', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->string('name')->unique();
-            $table->unsignedInteger('school_id')->nullable();
+            $table->unsignedBigInteger('school_id')->nullable();
             $table->foreign('school_id')->references('id')->on('schools')->onDelete('cascade')->onUpdate('cascade');
             $table->timestamps();
             $table->softDeletes();
